rename file OwnerData to types
