+++ workspaces/git-release-manager/plugins/git-release-manager/dev/mock/github/data.ts
@@ -1,4 +1,4 @@
-import { OwnerData } from './OwnerData';
+import { OwnerData } from './types';
 
 const DATA: Record<string, OwnerData> = {
   goodreleaser: {
